Rename Node cert tool to certgen
