+++ https/node/package.json
@@ -1,8 +1,8 @@
 {
   "type": "module",
   "scripts": {
-    "generate-certs": "node ../generate-https-certs-openssl-node/generate-ca.ts && node ../generate-https-certs-openssl-node/issue-cert.ts",
-    "install-root-ca": "node ../generate-https-certs-openssl-node/install-root-ca-cert.ts",
+    "generate-certs": "node ../certgen/generate-ca.ts && node ../certgen/issue-cert.ts",
+    "install-root-ca": "node ../certgen/install-root-ca-cert.ts",
     "start": "node https-server.ts",
     "test": "vitest run",
     "types": "tsc --noEmit"
